requirements.txt
fastapi
uvicorn
langchain
openai
psycopg2-binary
python-dotenv
sqlalchemy
pydantic
tqdm
requests
